+++ lib/utils/transform-response.ts
@@ -1,7 +1,7 @@
 import { IResult, IOptions } from '../interface';
 import { parseHeaders } from './parse-headers';
 
-export function transformData<T>({
+export function transformResponse<T>({
     response,
     options,
     xhr,
@@ -12,12 +12,12 @@ export function transformData<T>({
     xhr: XMLHttpRequest;
     statusField: string;
 }): T {
-    if (options && options.transformData) {
+    if (options && options.transformResponse) {
         const responseHeaders = xhr ? parseHeaders(xhr.getAllResponseHeaders()) : undefined;
         if (response && typeof response === 'object' && typeof response[statusField] !== 'undefined') {
-            (response as IResult).data = options.transformData((response as IResult).data, responseHeaders);
+            (response as IResult).data = options.transformResponse((response as IResult).data, responseHeaders);
         } else {
-            response = options.transformData(response, responseHeaders);
+            response = options.transformResponse(response, responseHeaders);
         }
     }
     return response;