+++ CHANGELOG.md
@@ -12,10 +12,10 @@
 
 # 3.4.0
 > 2020.08.11  
-1. <font color=blue>增强</font> 参数`options`添加`transformData`，自定义响应数据
+1. <font color=blue>增强</font> 参数`options`添加`transformResponse`，自定义响应数据
 ```
     /** 自定义响应数据 */
-    transformData?: (
+    transformResponse?: (
         /** 数据 */
         data: any,
         /** 响应头 */